Set CorrelationId as UUID if none set for request.
The `CorrelationId` allows tracking the relationship between a request
and a response. If a `CorrelationId` is not set, the authentication will
fail. To prevent this type of failure, the `CorrelationId` is now set
by default in the `Custom` constructor. It is still recommended to use
`setCorrelationId()` between each request to set a custom `CorrelationId`,
however this is not always necessary for all use cases.

Issue: https://github.com/makanaleu/wmt-marketplace-auth/issues/5
Signed-off-by: Kane McConnell <kane@makanal.eu>
CLA: f863b4f7bdee7d6997520ecaf537ddf0